[dagster-dbt] Support context in DbtCloudWorkspace.cli(..)
